Treat interactive calls to install differently than non-interactive calls.
When running as a service, cabbie does several things to handle new updates, including internally scheduling a reboot and triggering desktop notifications. These make less sense when running cabbie from the command line as a utility, so skip the reboot handling in those cases, in favor of simple print statements.

PiperOrigin-RevId: 347869524